Updated game scripts

+++ UnityAssets/Assets/Scripts/ScheduleOne/Variables/NumberVariable.cs
@@ -3,7 +3,7 @@ namespace ScheduleOne.Variables
 	public class NumberVariable : global::ScheduleOne.Variables.Variable<float>
 	{
 		public NumberVariable(string name, global::ScheduleOne.Variables.EVariableReplicationMode replicationMode, bool persistent, global::ScheduleOne.Variables.EVariableMode mode, global::ScheduleOne.PlayerScripts.Player owner, float value)
-			: base((string)null, default(global::ScheduleOne.Variables.EVariableReplicationMode), persistent: false, default(global::ScheduleOne.Variables.EVariableMode), (global::ScheduleOne.PlayerScripts.Player)null, (float)default)
+			: base((string)null, default(global::ScheduleOne.Variables.EVariableReplicationMode), persistent: false, default(global::ScheduleOne.Variables.EVariableMode), (global::ScheduleOne.PlayerScripts.Player)null, default)
 		{
 		}//IL_0024: Expected F4, but got O
